services/orchestrator/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
httpx>=0.28.0
google-genai>=1.0.0
